update api for getting visitedPlaces
